update store stats at starup
On the way refactor how sizes are store in the db
We store the size as integer in bytes in the db
but we retun the size as MiB to the user so it's more reabable

Reservation are done by MiB, so I also change the type from float to int

+++ store/jwt/reservation.go
@@ -18,7 +18,7 @@ type reservationClaims struct {
 	jwt.StandardClaims
 	ID           string // ID of the reservation
 	AdminID      string // ID of the user that has created the reservation, only him can renew
-	SizeReserved float64
+	SizeReserved uint64
 	Namespace    string
 }
